+++ CHANGELOG.md
@@ -10,6 +10,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Added
 
 - Support additional commands (#435):
+  - `BGREWRITEAOF` (#444)
   - `BGSAVE CANCEL` (#436)
   - `BGSAVE SCHEDULE` (#436)
   - `BGSAVE` (#436)
@@ -18,6 +19,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
   - `CLIENT TRACKING` (#451)
   - `CLIENT TRACKINGINFO` (#451)
   - `CLIENT UNPAUSE` (#437)
+  - `MEMORY DOCTOR` (#443)
+  - `MEMORY MALLOC-STATS` (#443)
+  - `MEMORY PURGE` (#443)
+  - `MEMORY STATS` (#443)
   - `SAVE` (#440)
 - Custom socket address resolution support via callback (#392)